done replicating python output
code still messy, quantization probably not supported

trace: {"region":"","traceId":"c4179cb3f587ce3871543ecc973b5066","userId":"xuyuheng.ken2","chatSessionId":"699eb41e6b1dccd57b6de217","replyToMessageId":"699eb4a17977a04fa1690d1b","messageId":"699eb4a16b1dccd57b6de23e","createTime":"2/25/2026, 4:36:49 PM","modelName":"gemini-3-pro","ckgStatus":"success","hasWorkspace":false,"hasFolder":false,"status":"completed","ppeEnvName":"","appName":"Trae CN - Local","isAutoMode":false,"callModel":"gemini-3-pro","isMaxMode":true,"agentProcessSupport":"v3"}

+++ rust/src/stft/istft.rs
@@ -25,7 +25,7 @@ pub fn istft(
     let mut output = vec![vec![0.0f32; reconstructed_len]; num_channels];
     let mut window_sum = vec![0.0f32; reconstructed_len];
 
-    let window_sq: Vec<f32> = plan.window.iter().map(|w| w * w).collect();
+    let window_sq: Vec<f32> = plan.window.iter().map(|w| (w * w) as f32).collect();
 
     for (ch, channel_spec) in spectrogram.iter().enumerate() {
         for frame_idx in 0..num_frames {
